+++ portalen/apps/cms/src/database/seed/data/component-pages.ts
@@ -1,4 +1,4 @@
-import type { Component } from "../..";
+import type { Component } from "../../../types";
 
 export const componentPagesSeed: Omit<Component, "id">[] = [
     {
